added authorization

+++ backend/routes/posts.js
@@ -3,6 +3,7 @@ const multer = require("multer");
 
 const Post = require("../models/post");
 const checkAuth = require("../middleware/check-auth");
+const { create } = require("../models/post");
 
 const router = express.Router();
 
@@ -37,7 +38,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -65,10 +67,15 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-      res.status(200).json({message: "Updata successfull!" });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Updata successfull!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" })
+      }
     });
   }
 );
@@ -108,10 +115,13 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId })
     .then(result => {
-      console.log(result);
-      res.status(200).json({ message: "Post deleted!" });
+      if (result.n > 0) {
+        res.status(200).json({ message: "Delete successfull!" });
+      } else {
+        res.status(401).json({ message: "Not authorized!" })
+      }
     });
 });
